[Pdm2-6] Extract functions as services
